modified:   .idea/workspace.xml 	modified:   app/src/main/AndroidManifest.xml 	new file:   app/src/main/java/cn/edu/ustc/appseed/clubseed/activity/AuthorActivity.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/activity/EventContentActivity.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/activity/MainActivity.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/activity/StartActivity.java 	new file:   app/src/main/java/cn/edu/ustc/appseed/clubseed/adapter/DrawerListViewAdapter.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/fragment/EventContentFragment.java 	deleted:    app/src/main/java/cn/edu/ustc/appseed/clubseed/fragment/EventShowFragment.java 	new file:   app/src/main/java/cn/edu/ustc/appseed/clubseed/fragment/NoticeFragment.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/fragment/SettingsFragment.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/jsondata/ListPhp.java 	modified:   app/src/main/java/cn/edu/ustc/appseed/clubseed/utils/AppUtils.java 	new file:   app/src/main/res/layout/activity_author.xml 	modified:   app/src/main/res/layout/activity_main.xml 	modified:   app/src/main/res/layout/activity_start.xml 	new file:   app/src/main/res/layout/drawer_list.xml 	new file:   app/src/main/res/menu/menu_author.xml 	new file:   app/src/main/res/menu/menu_main2.xml 	modified:   app/src/main/res/values/colors.xml 	new file:   app/src/main/res/values/fonts.xml 	modified:   app/src/main/res/values/strings.xml 	modified:   app/src/main/res/xml/preferences.xml

+++ app/src/main/java/cn/edu/ustc/appseed/clubseed/utils/AppUtils.java
@@ -2,6 +2,9 @@ package cn.edu.ustc.appseed.clubseed.utils;
 
 import android.app.Activity;
 import android.content.Context;
+import android.content.SharedPreferences;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.util.Log;
 
 import com.alibaba.fastjson.JSON;
@@ -21,11 +24,19 @@ import cn.edu.ustc.appseed.clubseed.jsondata.ListPhp;
  */
 public class AppUtils extends Activity{
     private static final String TAG = "JSON_FILE_ERROR";
-    private static final int URLSPLIT = 32;
     public static Context sAppContext;
+    public static SharedPreferences sSharedPreferences;
+    public static boolean isReadingMode;
+    public static ListPhp NullListPhp;
+    public static String version = "0.0.9";
+    public static String phoneInfo;
+
+    public static String changeFilename(String filename){
+        return filename.replace('/', '|');
+    }
 
     public static boolean saveCache(String filename, String content) {
-        filename = filename.substring(URLSPLIT);
+        filename = changeFilename(filename);
         FileOutputStream fos = null;
         byte[] buff = content.getBytes();
         try {
@@ -42,7 +53,7 @@ public class AppUtils extends Activity{
     }
 
     public static <T> T loadFile(String filename, Class<T> clazz) {
-        filename = filename.substring(URLSPLIT);
+        filename = changeFilename(filename);
         File f = new File(sAppContext.getCacheDir().getPath() + filename);
         byte[] buff = new byte[(int)f.length()];
         try {
@@ -57,7 +68,7 @@ public class AppUtils extends Activity{
     }
 
     public static boolean existCache(String filename){
-        filename = filename.substring(URLSPLIT);
+
         File file = new File(sAppContext.getCacheDir().getPath() + filename);
         Log.d("HR",file.exists()?filename+":EXIST":filename+":NOT EXIST");
         return file.exists();
@@ -69,6 +80,7 @@ public class AppUtils extends Activity{
      */
     public static boolean clearCache(){
         File[] files = sAppContext.getCacheDir().listFiles();
+        if(files==null)return false;
         for(File file : files){
             file.deleteOnExit();
         }
@@ -88,6 +100,21 @@ public class AppUtils extends Activity{
                     .build();
             String jsonString = client.newCall(listRequest).execute().body().string();
             return jsonString;
+        } catch (IOException e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    public static Bitmap getBitmapFromURL(String url){
+        final OkHttpClient client = new OkHttpClient();
+        try {
+            Request listRequest = new Request.Builder()
+                    .url(url)
+                    .build();
+            byte[] bytes = client.newCall(listRequest).execute().body().bytes();
+            Bitmap bitmap = BitmapFactory.decodeByteArray(bytes,0,bytes.length);
+            return bitmap;
         } catch (IOException e) {
             e.printStackTrace();
             return null;
